api-mock/requirements.txt
anyio==3.6.2
boto3==1.26.69
botocore==1.29.69
click==8.1.3
fastapi==0.91.0
h11==0.14.0
httptools==0.5.0
idna==3.4
jmespath==1.0.1
pydantic==1.10.4
python-dateutil==2.8.2
python-dotenv==0.21.1
python-magic==0.4.27
PyYAML==6.0
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
starlette==0.24.0
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
